service-workers/service-worker/fetch-request-xhr-sync-on-worker.https.html

<!DOCTYPE html>
<title>Service Worker: Synchronous XHR on Worker is intercepted</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

promise_test(async t => {
  const SCRIPT = 'resources/fetch-request-xhr-sync-on-worker-worker.js';
  const SCOPE = 'resources/fetch-request-xhr-sync-on-worker-scope/';
  const NON_EXISTENT_FILE = 'non-existent-file.txt';

  // In Chromium, the service worker scope matching for workers is based on
  // the URL of the parent HTML. So this test creates an iframe which is
  // controlled by the service worker first, and creates a worker from the
  // iframe.
  const registration = await service_worker_unregister_and_register(
    t,
    SCRIPT,
    SCOPE
  );
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(SCOPE + 'iframe_page');
  const result = await frame.contentWindow.performSyncXHROnWorker(
    NON_EXISTENT_FILE
  );

  t.add_cleanup(async () => {
    if (frame) frame.remove();
    if (registration) await registration.unregister();
  });

  assert_equals(
    result.status,
    200,
    'HTTP response status code for intercepted request'
  );
  assert_equals(
    result.responseText,
    'Response from service worker',
    'HTTP response text for intercepted request'
  );
}, 'Verify SyncXHR on Worker is intercepted');
</script>
